solana: refactor complete swap
